tests/wpt/tests/feature-policy/experimental-features/trust-token-redemption-supported-by-feature-policy.tentative.html

<!DOCTYPE html>
<title>Test that trust token redemption is advertised in the feature list</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
    assert_in_array('trust-token-redemption', document.featurePolicy.features());
  }, 'document.featurePolicy.features should advertise trust token redemption.');
</script>
